requirements.txt
-e .
pytest
aws_cdk.aws_sns
aws_cdk.aws_sns_subscriptions
wheel
